fix: PRレビュー指摘事項の修正
- douhashiハードコードをutils.GetGitHubRepoInfo()を使用して修正
  - cmd/utils.goのgetRepoIdentifier()関数を更新
  - GitHubリポジトリ情報から動的にowner/repo情報を取得するように変更

- Windows環境のビルドエラーをビルドタグで対応
  - daemon機能をOS別に分離（daemon_unix.go / daemon_windows.go）
  - 共通ロジックをdaemon_common.goに集約
  - Windows環境では機能制限があることを明確化
