+++ cmd/utils.go
@@ -1,20 +1,19 @@
 package cmd
 
 import (
+	"context"
 	"fmt"
 
-	"github.com/douhashi/osoba/internal/git"
+	"github.com/douhashi/osoba/internal/utils"
 )
 
 // getRepoIdentifier はリポジトリの識別子を取得します
 func getRepoIdentifier() (string, error) {
-	repoName, err := git.GetRepositoryName()
+	// GitHubリポジトリ情報を取得
+	repoInfo, err := utils.GetGitHubRepoInfo(context.Background())
 	if err != nil {
-		return "", fmt.Errorf("リポジトリ名の取得に失敗: %w", err)
+		return "", fmt.Errorf("リポジトリ情報の取得に失敗: %w", err)
 	}
 
-	// TODO: より正確な方法でオーナーを取得する
-	owner := "douhashi"
-
-	return fmt.Sprintf("%s-%s", owner, repoName), nil
+	return fmt.Sprintf("%s-%s", repoInfo.Owner, repoInfo.Repo), nil
 }
